BotTrainer/requirements.txt
google-genai>=1.0.0
streamlit>=1.42.0
kaggle>=1.7.0
python-dotenv>=1.0.0
scikit-learn>=1.6.0
pandas>=2.2.0
numpy>=2.0.0
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
